requirements.txt
fastapi==0.81.0
uvicorn == 0.18.3
line-bot-sdk == 2.3.0
aiohttp~=3.8.1
pydub~=0.25.1
ffmpeg~=1.4